# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 电力交易智能辅助决策系统 API
  description: |
    配套开发基线 5.5 API 目录总表。统一响应封装 ApiResponse；分页用 PageResult；
    鉴权 JWT Bearer；错误码分段见 README 第 4 条。所有路径省略网关前缀 /api/v1。
  version: 1.0.0
  license:
    name: 项目内部使用
    url: https://api.ptidss.example.com/license
  contact:
    name: PTIDSS 架构组
servers:
  - url: https://api.ptidss.example.com/api/v1
    description: 生产（示例域名）
  - url: https://api-test.ptidss.example.com/api/v1
    description: 测试
...
security:
  - bearerAuth: []
paths:
...
  /trade/rolling-plans:
    get:
      operationId: getTradeRollingPlans
      summary: 日滚动方案列表（多情景）
      tags: [trade]
      parameters:
        - { name: tradeDate, in: query, schema: { type: string, format: date } }
        - { name: scenario, in: query, schema: { type: string, enum: [baseline, conservative, aggressive] } }
        - { name: status, in: query, schema: { type: string, enum: [generated, confirmed, modified, rejected, executed] } }
      responses:
        '200':
          description: 方案列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/RollingPlan' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /trade/rolling-plans/{planId}/confirm:
    post:
      operationId: postTradeRollingPlansConfirm
      summary: 确认日滚动方案（进入申报）
      tags: [trade]
      parameters:
        - { name: planId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已确认
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /trade/declarations:
    post:
      operationId: postTradeDeclarations
      summary: 创建申报单（含合规预检）
      tags: [trade]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tradeDate, marketType, stage, items]
              properties:
                tradeDate: { type: string, format: date }
                marketType: { type: string, enum: [intra_province, inter_province] }
                stage: { type: string, enum: [day_ahead, real_time, rolling] }
                items: { type: array, items: { type: object }, description: 申报段（时段/量/价） }
                sourcePlanId: { type: string, description: 来源方案/会话 }
      responses:
        '200':
          description: 创建成功（含合规校验结果）
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: object
                        properties:
                          declarationId: { type: string }
                          complianceCheck: { type: object, description: 段数/限价/持仓校验结果 }
        '4XX': { $ref: '#/components/responses/BizError' }
    get:
      operationId: getTradeDeclarations
      summary: 申报单列表（状态/日期筛选）
      tags: [trade]
      parameters:
        - { name: tradeDate, in: query, schema: { type: string, format: date } }
        - { name: status, in: query, schema: { type: string, enum: [draft, pending_submit, submitted, receipted, partially_matched] } }
        - { $ref: '#/components/parameters/PageParam' }
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { $ref: '#/components/schemas/PageResult' } }
        '4XX': { $ref: '#/components/responses/BizError' }
  /trade/declarations/{id}/submit:
    post:
      operationId: postTradeDeclarationsSubmit
      summary: 提交申报至交易中心
      tags: [trade]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 提交成功（回执号）
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: object
                        properties: { receiptNo: { type: string } }
        '4XX': { $ref: '#/components/responses/BizError' }
  /trade/results:
    get:
      operationId: getTradeResults
      summary: 成交结果查询（按日/市场）
      tags: [trade]
      parameters:
        - { name: tradeDate, in: query, required: true, schema: { type: string, format: date } }
        - { name: marketType, in: query, schema: { type: string, enum: [intra_province, inter_province] } }
      responses:
        '200':
          description: 成交明细
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/TradeResult' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /trade/positions:
    get:
      operationId: getTradePositions
      summary: 持仓曲线（中长期+现货合成视图）
      tags: [trade]
      parameters:
        - { name: tradeDate, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          description: 持仓曲线
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: object
                        properties:
                          longTerm: { type: array, items: { type: number }, description: 96 点中长期持仓 }
                          spot: { type: array, items: { type: number } }
                          net: { type: array, items: { type: number } }
        '4XX': { $ref: '#/components/responses/BizError' }

  # ==================== settlement / ocr ====================
...
  /trade/gateway/config:
    get:
      operationId: getTradeGatewayConfig
      summary: 交易网关配置读取（含加密连接参数脱敏视图；仅 admin 角色，V3G-03 收紧）
      tags: [trade]
      responses:
        '200':
          description: 配置
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { type: object } }
        '4XX': { $ref: '#/components/responses/BizError' }
    put:
      operationId: putTradeGatewayConfig
      summary: 交易网关配置写入（appKey/appSecret 加密落库；仅 admin 角色，V3G-03 收紧）
      tags: [trade]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                host: { type: string }
                appKey: { type: string }
                appSecret: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /trade/gateway/test:
    post:
      operationId: postTradeGatewayTest
      summary: 网关连通性测试（解密密钥发起外部请求；仅 admin 角色，V3G-04 收紧）
      tags: [trade]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { host: { type: string } }
      responses:
        '200':
          description: 测试结果
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { type: object } }
        '4XX': { $ref: '#/components/responses/BizError' }

  # ==================== admin 配置中心（V3.0，33 项内置） ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: pageNo
      in: query
      description: 分页参数（pageNo/pageSize，对齐后端 @RequestParam 实现）
      schema:
        type: object
        properties:
          pageNo: { type: integer, minimum: 1, default: 1 }
          pageSize: { type: integer, maximum: 200, default: 20 }
  responses:
    BizError:
      description: 业务错误（错误码分段见 README）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: { type: integer, description: 0 成功；非 0 见错误码分段表, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true, description: 业务数据（随接口变化） }
        traceId: { type: string, description: 全链路追踪 ID }
    ErrorResponse:
      type: object
      properties:
        code: { type: integer, example: 14001 }
        message: { type: string }
        traceId: { type: string }
        detail: { type: object, nullable: true, description: 参数错误明细 }
    PageResult:
      type: object
      properties:
        list: { type: array, items: { type: object } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    RollingPlan:
      type: object
      properties:
        planId: { type: string }
        tradeDate: { type: string, format: date }
        scenario: { type: string }
        expectedRevenue: { type: number }
        riskMetrics: { type: object }
        status: { type: string }
        createdBy: { type: string }
    TradeResult:
      type: object
      properties:
        declarationId: { type: string }
        matchedVolume: { type: number }
        matchedPrice: { type: number }
        matchedCurve: { type: array, items: { type: number } }
        settlementImpact: { type: object }
...